refac: centralize my custom types to `manifold_types.py`
